requirements.txt

requests
psycopg2-binary
textblob
streamlit
pandas
